# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TutorConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List tuition posts with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: class
          in: query
          schema:
            type: string
        - name: subject
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: mode
          in: query
          schema:
            type: string
            enum: [home, online]
        - name: medium
          in: query
          schema:
            type: string
            enum: [english, hindi]
        - name: minFee
          in: query
          schema:
            type: integer
        - name: maxFee
          in: query
          schema:
            type: integer
        - name: genderPreference
          in: query
          schema:
            type: string
            enum: [male, female, any]
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"

    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new tuition post (parent only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TuitionPost"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a tuition post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TuitionPost"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updatePost
      tags: [posts]
      summary: Update a tuition post (owner or admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TuitionPost"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a tuition post (owner or admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts/my:
    get:
      operationId: getMyPosts
      tags: [posts]
      summary: Get current user's posts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TuitionPost"
  /posts/{id}/contact:
    get:
      operationId: getPostContact
      tags: [posts]
      summary: Get contact details (requires payment)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "402":
          description: Payment required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    CreatePostRequest:
      type: object
      properties:
        title:
          type: string
        class:
          type: string
        subjects:
          type: string
        genderPreference:
          type: string
          enum: [male, female, any]
          nullable: true
        medium:
          type: string
          enum: [english, hindi]
        mode:
          type: string
          enum: [home, online]
        address:
          type: string
        duration:
          type: number
        daysPerWeek:
          type: integer
        monthlyFee:
          type: integer
        description:
          type: string
          nullable: true
        contactPhone:
          type: string
        state:
          type: string
      required:
        - title
        - class
        - subjects
        - medium
        - mode
        - address
        - duration
        - daysPerWeek
        - monthlyFee
        - contactPhone
        - state
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
        class:
          type: string
        subjects:
          type: string
        genderPreference:
          type: string
          enum: [male, female, any]
          nullable: true
        medium:
          type: string
          enum: [english, hindi]
        mode:
          type: string
          enum: [home, online]
        address:
          type: string
        duration:
          type: number
        daysPerWeek:
          type: integer
        monthlyFee:
          type: integer
        description:
          type: string
          nullable: true
        contactPhone:
          type: string
        state:
          type: string
        featured:
          type: boolean
    TuitionPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        class:
          type: string
        subjects:
          type: string
        genderPreference:
          type: string
          enum: [male, female, any]
          nullable: true
        medium:
          type: string
          enum: [english, hindi]
        mode:
          type: string
          enum: [home, online]
        address:
          type: string
        duration:
          type: number
        daysPerWeek:
          type: integer
        monthlyFee:
          type: integer
        state:
          type: string
        description:
          type: string
          nullable: true
        featured:
          type: boolean
        createdById:
          type: integer
        createdByName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - class
        - subjects
        - medium
        - mode
        - address
        - duration
        - daysPerWeek
        - monthlyFee
        - state
        - featured
        - createdById
        - createdAt
    PostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/TuitionPost"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - posts
        - total
        - page
        - totalPages
    ContactResponse:
      type: object
      properties:
        phone:
          type: string
      required:
        - phone
...